requirements.txt

fastapi==0.115.12
uvicorn==0.34.3
starlette==0.46.2
python-dotenv==1.1.0


pydantic==2.11.5
pydantic_core==2.33.2
annotated-types==0.7.0
typing-inspection==0.4.1
typing_extensions==4.14.0


anyio==4.9.0
h11==0.16.0
httpx==0.28.1
httpcore==1.0.9
sniffio==1.3.1
websockets==15.0.1
httptools==0.6.4


click==8.2.1
colorama==0.4.6
watchfiles==1.0.5
exceptiongroup==1.3.0


SQLAlchemy==2.0.41
psycopg2-binary==2.9.10


pandas
numpy
matplotlib
seaborn
yfinance
mplfinance


beautifulsoup4==4.13.4
soupsieve==2.7
requests==2.32.4
PyYAML==6.0.2
charset-normalizer==3.4.2
urllib3==2.4.0
idna==3.10


certifi==2025.4.26
